pc_host/main.py: Allow numbers in DUT names
The line that is changed gets a device name from /etc/daft/devices.cfg
sections eg. [Minnowboard1] and removes all the numbers from the name.
Change it that it only removes numbers from the end of the name and also
remove underscores. This way a name like [Joule_570x_1] can be used.

Signed-off-by: Simo Kuusela <simo.kuusela@intel.com>