initialize_testing_harness: Separate the service to two parts
Make new start_libcomposite.service that handless starting the USB
emulation and leave other things to initialize_testing_harness.service.
This way stopping and starting the USB emulation is cleaner. Also this
allows developing the USB emulation more by allowing for example
different arguments or a settings file for it.

Signed-off-by: Simo Kuusela <simo.kuusela@intel.com>