README.MD: NTP server installation
Add section to cover installing NTP server. This NTP server is installed on PC_HOST and client installed on the BBBs. The NTP is necessary to correct the time in Beaglebones (because it does not have a inner battery). Beaglebone polls the NTP server everytime it is booted changing the time to correct one. This also fixes the problem with incorrect timestamps in aft log files.

Signed-off-by: Christian da Costa <christian.da.costa@intel.com>