testing_harness: Set beaglebones values to default prior testing
Beaglebone runs a script called initialize_testing_harness each time it boots. This particular script sets some values such as GPIO pins  etc. to default. The problems is that, if a test changes these values, they are not be set back to default until the Beaglebone reboots and hence might fail other tests that are run after. This fix resets the Beaglebones GPIO20 pin back to default value to avoid this.

Signed-off-by: Christian da Costa <christian.da.costa@intel.com>