testing_harness: Add Dropbear support for writing sshkeys when flashing
When flashing an image with DAFT, the system mounts the flashed filesystem and writes correct sshkeys to the authorized_key folder so it can later ssh to it to run tests. However, if the flashed image uses DropBear instead of OpenSSH, it uses a symbolic link in the .ssh/authorized_keys, which breaks when mounting the filesystem. Updated the system to first try adding the sshkeys into the standard .ssh/authorized keys file, and if it fails, directly to dropbears own authorized_keys file.

Signed-off-by: Christian da Costa <christian.da.costa@intel.com>